renamed set_custom_priorities.py to set_custom_overrides.py set_custom_overrides will override all metadata types (provided overrides are defined) --reset flag still only applies to priorities changed overall structure of how overrides are stored: STUDY: {METADATA_TYPE1: {ATTRIBUTE1: VALUE1, ...}, ...}
